modify the cost_update to accept entire cost_table as an argument
