+++ lab5/src/sender.c
@@ -3,7 +3,7 @@
 #include <string.h>
 #include <sys/socket.h>
 #include <unistd.h>
-#include "cost.h"
+#include "costs.h"
 #include "sender.h"
 #include "logger.h"
 #include "machine.h"
@@ -20,7 +20,7 @@ void read_changes(Config* cfg) {
 
         int target;
         int new_cost;
-        printf("Enter <id> <new_cost>: ");
+        printf("Enter <id> <new_cost>(Current Machine: %d):  ", cfg->machine->id);
         fflush(stdout);
 
         scanf("%d %d", &target, &new_cost);
